termios: extract into module/termios/ following the _socket pattern
importing.rs shrinks by 658 lines.
